Try it with a cinderclient
Ok, straight to importing the code in grpc might be a bit
ambitious, there are a number of crappy things we have to
deal with there around Cinder's object model and a good deal
of other work around our process forks etc.

There's some work going on around this independently with
cinderlib, so maybe revisit if this seems feasible.

To test feasibility let's back up to the simpler case and
use cinderlcient.  If we get this working then move to the
next step of introducing a new API service in Cinder, then
head down the path of just *using* Cinder code without services
if it seems interesting.

Personally I prefer just writing things in golang at this point,
but it's interesting to see what can be done here.