Implement a basic create and delete
This adds basic create/delete functionality and a start at
figuring out List.  The trick with list is figuring out how
to add the Entries into the ListVolumesResponse object.

Basic create/delete work and the README is updated with some
notes on how to run this using the Rexray CSC tool.