Add an example of importing Cinder directly
Ideally we wouldn't need an cidner-api service, or the cinderlcient.
This patch adds a version that just imports the cinder.volume code
directly and makes calls directly using the python objects.

The new file is *direct_server.py* and externally works the same as
the *csi_server.py* file, the main difference is that it doesn't
need/use keystone or the Cinder REST API Server.

Something like this could be cleaned up, dropped into Cinder and
enabled via configuration as an option for connecting Cinder to
CSI consumers.